requirements.txt
asgiref == 3.6.0
Django == 4.0
django-etc == 1.4.0
django-hitcount == 1.3.5
django-modeltranslation == 0.18.9
Pillow == 9.4.0
python-decouple == 3.8 
python-gettext == 4.1
sqlparse == 0.4.4
typing-extensions == 4.5.0




d